requirements.txt
# Core
aiosqlite>=0.19.0
thefuzz[speedup]>=0.22.0
pydantic>=2.0
pydantic-settings>=2.0
structlog>=24.0
python-dotenv>=1.0

# Telegram
python-telegram-bot>=21.0

# Groq
groq>=0.9.0
tenacity>=8.2.0

# Google Calendar
google-auth>=2.0
google-api-python-client>=2.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
